+++ VideoPlayer/VideoPlayer.js
@@ -845,7 +845,7 @@ const VideoPlayerBase = class extends Component {
 
 	componentDidMount () {
 		on('mousemove', this.activityDetected);
-		if (platform.touch) {
+		if (platform.touchEvent) {
 			on('touchmove', this.activityDetected);
 		}
 		document.addEventListener('keydown', this.handleGlobalKeyDown, {capture: true});
@@ -943,7 +943,7 @@ const VideoPlayerBase = class extends Component {
 
 	componentWillUnmount () {
 		off('mousemove', this.activityDetected);
-		if (platform.touch) {
+		if (platform.touchEvent) {
 			off('touchmove', this.activityDetected);
 		}
 		document.removeEventListener('keydown', this.handleGlobalKeyDown, {capture: true});
@@ -1697,7 +1697,7 @@ const VideoPlayerBase = class extends Component {
 		this.playbackRate = String(rate);
 		const pbNumber = calcNumberValueOfPlaybackRate(this.playbackRate);
 
-		if (!platform.webos) {
+		if (platform.type !== 'webos') {
 			// ReactDOM throws error for setting negative value for playbackRate
 			this.video.playbackRate = pbNumber < 0 ? 0 : pbNumber;