+++ tests/ui/specs/PopupTabLayout/PopupTabLayout-specs.js
@@ -229,6 +229,15 @@ describe('PopupTabLayout', function () {
 					expect(actual).to.equal(expected);
 				});
 
+				it('should change the tab content immediately when pressing the enter key on the tab menu (WRP-26238)', async function () {
+					// Press the down key and then enter key to go to the second panel
+					await Page.spotlightDown();
+					await Page.delay(50);
+					await Page.spotlightSelect();
+
+					// Make sure the focus is on the second panel
+					expect(await browser.execute(getFocusedText)).to.equal('Advanced Audio');
+				});
 			});
 
 			// Note: To verify that testing is not possible.