+++ CHANGELOG.md
@@ -2,6 +2,58 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Added
+
+- `sandstone/Scroller.ContentContainerDecorator` to allow component to be a scroll target when its descendant is focused
+
+### Fixed
+
+- `sandstone/VirtualList` to show the focused item properly when switching from pointer mode to 5-way mode
+- `sandstone/VirtualList.VirtualGridList` to show the focused item fully when focus moves via 5-way keys inside a list
+
+## [2.8.0] - 2024-02.07
+
+### Fixed
+
+- `sandstone/Scroller` with `editable` prop to remain focused on the selected item when completing edit by down or enter key in pointer mode
+- `sandstone/Scroller` not to forward `onBack` handler of `sandstone/Panels` when focus moves from scroll thumb to BodyText via back key
+- `sandstone/VirtualList` to have proper scroll position when item with affordance is larger than scroll area
+
+## [2.5.12] - 2024-02-02
+
+### Fixed
+
+- `sandstone/WizardPanels` to read out the correct step when using `current` prop
+
+## [2.7.13] - 2023-12-08
+
+### Changed
+
+- `sandstone/Dropdown` to read out more details
+- `sandstone/Scroller` with `editable` prop to complete editing when 'down' key is pressed during editing
+- `sandstone/Slider` to read out more details
+- `sandstone/TabLayout` to move focus from tab contents to tab menu via back key
+- `sandstone/TabLayout` to read out more details
+
+### Fixed
+
+- `sandstone/Scroller` with `editable` prop to rearrange items properly when moving pointer very fast
+- `sandstone/Scroller` with `editable` prop to complete editing when focus left by 5-way key in pointer mode
+
+## [2.7.12] - 2023-10-23
+
+### Fixed
+
+- `sandstone/Scroller` to read out the announcement of completion properly when `editable` is given
+
+## [2.7.11] - 2023-10-13
+
+### Fixed
+
+- `sandstone/TabLayout` to revert 2.7.9 fix that change back key behavior
+
 ## [2.7.10] - 2023-09-20
 
 ### Fixed
@@ -18,7 +70,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 ### Fixed
 
 - `sandstone/Panels.Header` to not show `slotAfter` in incorrect position at first rendering when `centered` is given
-- `sandstone/Scroller` to read out properly when `editable` is given 
+- `sandstone/Scroller` to read out properly when `editable` is given
 
 ## [2.7.8] - 2023-08-31